+++ src/components/panel.js
@@ -79,6 +79,8 @@ module.exports = function(app) {
           '</div>' +
           '<form-builder-option property="customClass"></form-builder-option>' +
           '<form-builder-option property="hideLabel"></form-builder-option>' +
+          '<form-builder-option property="collapsible"></form-builder-option>' +
+          '<form-builder-option property="collapsed" ng-if="component.collapsible"></form-builder-option>' +
           '<form-builder-option property="tableView"></form-builder-option>' +
         '</ng-form>'
       );