+++ src/constants/commonOptions.js
@@ -347,5 +347,15 @@ module.exports = {
     label: 'Initial Focus',
     type: 'checkbox',
     tooltip: 'Make this field the initially focused element on this form.'
+  },
+  'collapsible': {
+    label: 'Collapsible',
+    type: 'checkbox',
+    tooltip: 'It\'s possible to hide body clicking the title.'
+  },
+  'collapsed': {
+    label: 'Collapsed',
+    type: 'checkbox',
+    tooltip: 'Component\'s body initially hidden.'
   }
 };